IRectRelation, IRectMaskRelation refactor 2nd (massive)
IRectMaskRelation further renamed to IRectMaskRelationInline.

IRectMaskRelation now refers to a non-inline interface with a TestMaybe method. The mask type is still a required generic type parameter.

A non-generic version of IRectMaskRelation is also created. The TestMaybe method accepts IRectMaskArith. However, the implementer of this interface may have to do a lot of work to actually invoke the underlying bit testing methods.